+++ NtoLib/Recipes/MbeTable/ModulePresentation/DataAccess/TargetItemsProvider.cs
@@ -17,18 +17,18 @@ public sealed class TargetItemsProvider : IComboBoxItemsProvider
         _context = context;
     }
 
-    public List<KeyValuePair<int, string>> GetItems(int rowIndex, ColumnIdentifier columnKey)
+    public List<KeyValuePair<short, string>> GetItems(int rowIndex, ColumnIdentifier columnKey)
     {
         var vm = _context.GetStepViewModel(rowIndex);
         if (vm == null)
-            return new List<KeyValuePair<int, string>>();
+            return new List<KeyValuePair<short, string>>();
 
         var result = vm.GetComboItems(columnKey);
         if (result.IsFailed || result.Value == null)
-            return new List<KeyValuePair<int, string>>();
+            return new List<KeyValuePair<short, string>>();
 
         return result.Value
-            .Select(kv => new KeyValuePair<int, string>(kv.Key, kv.Value))
+            .Select(kv => new KeyValuePair<short, string>(kv.Key, kv.Value))
             .ToList();
     }
 }
